Tests/LibWeb/Text/expected/scrollend-after-wheel-scroll.txt

scrollend target: scroller
scrollend bubbles: false
scroll offset: 0,80
scroller scrollend count: 1
document scrollend count: 0
